Error Handling Inside express and Outside express

+++ services/categoryService.js
@@ -1,6 +1,8 @@
 const Category = require("../models/categoryModel");
 const slugify = require("slugify");
 const asyncHandler = require("express-async-handler");
+const mongoose = require("mongoose");
+const ApiError =require("../utils/ApiError");
 //@desc Get all Categories
 //@route GET /api/v1/categories
 //@access Public
@@ -9,21 +11,21 @@ exports.getCategories = asyncHandler(async (req, res) => {
   const limit = req.query.limit * 1 || 5;
   const skip = (page - 1) * limit;
   const categories = await Category.find({}).skip(skip).limit(limit);
-  res.status(200).json({ results: categories.length, page, data: categories });
-  res.status(400).json({
-    message: err.message,
-  });
+  res.status(200).json({ status: 'success',results: categories.length, page, data: categories });
 });
 //@desc Get Category by ID
 //@route GET /api/v1/categories/:id
 //@access Public
-exports.getCategory = asyncHandler(async (req, res) => {
+exports.getCategory = asyncHandler(async (req, res,next) => {
   const id = req.params.id;
+  if(!mongoose.Types.ObjectId.isValid(id)){
+    return next(new ApiError(404,`category not found with id ${id}`));
+  }
   const category = await Category.findById(id);
   if (!category) {
-    res.status(404).json({ message: "Category not found" });
+    return next(new ApiError(404,`category not found with id ${id}`));
   }
-  res.status(200).json({ data: category });
+  res.status(200).json({ status: 'success' ,data: category });
 });
 
 //@desc Create Category
@@ -35,17 +37,17 @@ exports.createCategory = asyncHandler(async (req, res) => {
     name,
     slug: slugify(name),
   });
-  res.status(201).json({ data: category });
-  res.status(400).json({
-    message: err.message,
-  });
+  res.status(201).json({ status: 'success',data: category  });
 });
 
 //@desc Update Category by ID
 //@route PUT /api/v1/categories/:id
 //@access Private
-exports.updateCategory = asyncHandler(async (req, res) => {
+exports.updateCategory = asyncHandler(async (req, res,next) => {
   const id = req.params.id;
+  if(!mongoose.Types.ObjectId.isValid(id)){
+    return next(new ApiError(400,`category not found with id ${id}`));
+  }
   const name = req.body.name;
   const category = await Category.findOneAndUpdate(
     { _id: id },
@@ -53,19 +55,22 @@ exports.updateCategory = asyncHandler(async (req, res) => {
     { new: true }
   );
   if (!category) {
-    res.status(404).json({ message: "Category not found" });
+    return next(new ApiError(404,`category not found with id ${id}`));
   }
-  res.status(200).json({ data: category });
+  res.status(200).json({status: 'success',data: category });
 });
 //@desc Delete Category by ID
 //@route DELETE /api/v1/categories/:id
 //@access Private
 
-exports.deleteCategory = asyncHandler(async (req, res) => {
+exports.deleteCategory = asyncHandler(async (req, res,next) => {
   const id = req.params.id;
+  if(!mongoose.Types.ObjectId.isValid(id)){
+    return next(new ApiError(400,`category not found with id ${id}`));
+  }
   const category = await Category.findByIdAndDelete(id);
   if (!category) {
-    res.status(404).json({ message: "Category not found" });
+    return next(new ApiError(404,`category not found with id ${id}`));
   }
-  res.status(200).json({ message: "Category deleted successfully" });
+  res.status(200).json({ status: 'success',message: "Category deleted successfully" });
 });
